move parse_yaml_msg_id_file() definition to generate_microRTPS_bridge.py
